pyproject.toml
[project]
name = "download-seedrcc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "aioseedrcc>=0.3.2",
    "docopt>=0.6.2",
    "ipython>=9.0.2",
    "python-dotenv>=1.0.1",
    "seedr>=1.2.3",
    "seedrapi>=1.0",
    "tqdm>=4.67.1",
]
